use more specific traceables for ApiCalls

+++ app/lib/graph/entity_extractor.rb
@@ -1,14 +1,26 @@
 module Graph
   class EntityExtractor
-    def initialize(model_config, traceable: nil)
+    def initialize(model_config)
       @model_config = model_config
-      @traceable = traceable
       @prompt_template = ERB.new(Graph::Prompts::ENTITY_EXTRACTION_PROMPT)
     end
 
     def extract(chunk)
+      if chunk.content.size < minimum_chunk_size
+        Rails.logger.info(
+          "Chunk #{chunk.id} is too small to extract entities (#{chunk.content.size} < #{minimum_chunk_size})"
+        )
+        return
+      end
+
       chunk.graph_entity_descriptions.destroy_all
 
+      extract_chunk_entities(chunk)
+    end
+
+    private
+
+    def extract_chunk_entities(chunk)
       entities_for(chunk).each_line do |line|
         break if line =~ /########/
 
@@ -23,8 +35,6 @@ module Graph
         end
       end
     end
-
-    private
 
     def handle_entity(values, chunk)
       unless entity_types.include?(values[:subtype])
@@ -60,7 +70,7 @@ module Graph
     end
 
     def entities_for(chunk)
-      entities = client.complete(prompt_for(chunk.content))
+      entities = client.complete(prompt_for(chunk.content), traceable: chunk)
 
       Rails.logger.debug { "Extracted entities:\n#{entities}" }
 
@@ -81,13 +91,16 @@ module Graph
     end
 
     def prompt_for(input_text)
-      @prompt_template.result(binding)
+      prompt = @prompt_template.result(binding)
+      Rails.logger.info("Sending completion request:\n#{prompt}")
+
+      prompt
     end
 
     ###
     # template values
     def tuple_delimiter
-      "|"
+      " | "
     end
 
     def record_delimiter
@@ -99,11 +112,15 @@ module Graph
     end
 
     def entity_types
-      %w[organization person event concept technology mission location role project]
+      %w[organization person event concept technology mission location role project].join(" ")
     end
 
     def input_text
       @chunk.content
+    end
+
+    def minimum_chunk_size
+      @minimum_chunk_size ||= Setting.get("minimum_chunk_size_for_extraction", default: 128)
     end
   end
 end
